# dsa-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for the DSA Backend application.",
        "title": "DSA Backend API",
        "contact": {},
        "version": "0.0.1"
    },
    "host": "localhost:8000",
    "basePath": "/api",
    "paths": {
...
        "/problem/fetch/detail/{lectureid}/{problemid}": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": [
                            "me"
                        ]
                    }
                ],
                "description": "Get detailed information about a specific problem within a lecture.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fetch"
                ],
                "summary": "Get problem detail",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Lecture ID",
                        "name": "lectureid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Problem ID",
                        "name": "problemid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.ProblemDetail"
                        }
                    },
                    "400": {
                        "description": "invalid request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "404": {
                        "description": "problem not found",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "failed to get problem",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/problem/fetch/list": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": [
                            "me"
                        ]
                    }
                ],
                "description": "get all lecture entries, each containing its problem entries. When you don't have scopes \"grading\" or \"admin\", you will only see lecture entries that are published.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fetch"
                ],
                "summary": "list all problem entry, nested in lecture entry.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/util.LectureEntry"
                            }
                        }
                    },
                    "500": {
                        "description": "failed to get lecture entries",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/problem/fetch/requiredfiles": {
            "get": {
                "security": [
                    {
                        "OAuth2Password": [
                            "me"
                        ]
                    }
                ],
                "description": "Get a list of required files for each lecture, including problem-specific files and a report template.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fetch"
                ],
                "summary": "List required files for each lecture",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/problem.ListRequiredFilesResponse"
                        }
                    },
                    "500": {
                        "description": "failed to get lecture list",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "problem.ListRequiredFilesResponse": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/problem.RequiredFiles"
                    }
                }
            }
        },
...
        "problem.RequiredFiles": {
            "type": "object",
            "properties": {
                "files": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "lecture_id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "response.Error": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "object",
                    "properties": {
                        "body": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "util.FileData": {
            "type": "object",
            "properties": {
                "compression": {
                    "description": "e.g., \"gzip\"",
                    "type": "string"
                },
                "data": {
                    "description": "base64 encoded data, compressed by gzip",
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "original_size": {
                    "type": "integer"
                }
            }
        },
        "util.LectureEntry": {
            "type": "object",
            "properties": {
                "deadline": {
                    "type": "integer"
                },
                "lecture_id": {
                    "type": "integer"
                },
                "problems": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/util.ProblemEntry"
                    }
                },
                "start_date": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "util.ProblemDetail": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "lecture_id": {
                    "type": "integer"
                },
                "memory_mb": {
                    "type": "integer"
                },
                "problem_id": {
                    "type": "integer"
                },
                "required_files": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "test_files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/util.FileData"
                    }
                },
                "time_ms": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "util.ProblemEntry": {
            "type": "object",
            "properties": {
                "lecture_id": {
                    "type": "integer"
                },
                "problem_id": {
                    "type": "integer"
                },
                "registered_at": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "/api/user/login",
            "scopes": {
                "admin": "Grants any rights related to admin user",
                "manager": "Grants any rights related to manager user",
                "me": "Grants any rights related to the current user"
            }
        }
    }
}